requirements.txt
trulens_eval
openai
python-dotenv
chromadb
llama-index
IPython
boto3
botocore